Cambio datos primitivos y asientos de vuelos a avion

+++ src/main/java/com/example/aereopuerto/Specifications/AvionSpecification.java
@@ -8,55 +8,60 @@ public class AvionSpecification {
 
     public static Specification<Avion> porIdentificador(String identificador) {
         return (root, query, criteriaBuilder) -> {
-            if (identificador == null || identificador.trim().isEmpty()) return null;
+            if (identificador == null || identificador.trim().isEmpty())
+                return null;
             return criteriaBuilder.like(
                     criteriaBuilder.lower(root.get("identificador")),
-                    "%" + identificador.toLowerCase() + "%"
-            );
+                    "%" + identificador.toLowerCase() + "%");
         };
     }
 
     public static Specification<Avion> porModelo(String modelo) {
         return (root, query, criteriaBuilder) -> {
-            if (modelo == null || modelo.trim().isEmpty()) return null;
+            if (modelo == null || modelo.trim().isEmpty())
+                return null;
             return criteriaBuilder.like(
                     criteriaBuilder.lower(root.get("modelo")),
-                    "%" + modelo.toLowerCase() + "%"
-            );
+                    "%" + modelo.toLowerCase() + "%");
         };
     }
 
     public static Specification<Avion> porEstado(estadoAvion estado) {
         return (root, query, criteriaBuilder) -> {
-            if (estado == null) return null;
+            if (estado == null)
+                return null;
             return criteriaBuilder.equal(root.get("estado"), estado);
         };
     }
 
     public static Specification<Avion> capacidadPasajerosMinimaDesde(Integer minimo) {
         return (root, query, criteriaBuilder) -> {
-            if (minimo == null) return null;
+            if (minimo == null)
+                return null;
             return criteriaBuilder.greaterThanOrEqualTo(root.get("capacidadPasajeros"), minimo);
         };
     }
 
     public static Specification<Avion> capacidadPasajerosMaximaHasta(Integer maximo) {
         return (root, query, criteriaBuilder) -> {
-            if (maximo == null) return null;
+            if (maximo == null)
+                return null;
             return criteriaBuilder.lessThanOrEqualTo(root.get("capacidadPasajeros"), maximo);
         };
     }
 
-    public static Specification<Avion> capacidadBodegaMinimaDesde(Float minimo) {
+    public static Specification<Avion> capacidadBodegaMinimaDesde(Double minimo) {
         return (root, query, criteriaBuilder) -> {
-            if (minimo == null) return null;
+            if (minimo == null)
+                return null;
             return criteriaBuilder.greaterThanOrEqualTo(root.get("capacidadBodega"), minimo);
         };
     }
 
-    public static Specification<Avion> capacidadBodegaMaximaHasta(Float maximo) {
+    public static Specification<Avion> capacidadBodegaMaximaHasta(Double maximo) {
         return (root, query, criteriaBuilder) -> {
-            if (maximo == null) return null;
+            if (maximo == null)
+                return null;
             return criteriaBuilder.lessThanOrEqualTo(root.get("capacidadBodega"), maximo);
         };
     }
