+++ src/main/java/com/example/aereopuerto/dto/AvionDTO.java
@@ -3,11 +3,10 @@ package com.example.aereopuerto.dto;
 import com.example.aereopuerto.model.enums.estadoAvion;
 import lombok.Data;
 
-
 @Data
 public class AvionDTO {
     private String identificador;
-    private Float capacidadBodega;
+    private Double capacidadBodega;
     private Integer capacidadPasajeros;
     private String modelo;
     private estadoAvion estado;
