+++ src/main/java/com/example/aereopuerto/model/Avion.java
@@ -19,25 +19,23 @@ public class Avion implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="avion_id")
+    @Column(name = "avion_id")
     private Integer id;
 
     @Column(nullable = false)
     private String identificador;
 
-    @Column (nullable = false, name= "capacidad_bodega")
-    private Float capacidadBodega;
+    @Column(nullable = false, name = "capacidad_bodega")
+    private Double capacidadBodega;
 
-    @Column (nullable = false, name= "capacidad_pasajeros")
+    @Column(nullable = false, name = "capacidad_pasajeros")
     private Integer capacidadPasajeros;
 
-    @Column (nullable = false)
+    @Column(nullable = false)
     private String modelo;
 
-
-    @Column (nullable = false)
+    @Column(nullable = false)
     @Enumerated(EnumType.STRING)
     private estadoAvion estado;
-
 
 }
